selfdev-agency/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.8
pydantic==2.10.6
uvicorn==0.34.0
# httpx==0.28.1
httpx
waitress==3.0.2
redis==5.2.1
langchain==0.3.18
# langchain-core==0.3.33
langchain-core
langchain-community==0.3.17
langchain-text-splitters==0.3.6
langchain-google-community[drive]==2.0.4
langchain-googledrive==0.3.35
langchain-chroma==0.2.6
langchain-unstructured==0.1.6
langgraph==0.2.72
tiktoken==0.9.0
unstructured==0.16.20
unstructured-client==0.30.0
unstructured[all-docs]==0.16.20
faiss-cpu==1.10.0
google-api-python-client==2.161.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-magic==0.4.27
chromadb==1.0.20
weaviate-client==4.10.4
langchain-weaviate==0.0.4
libmagic==1.0

langchain[groq]==0.3.18
langchain[openai]==0.3.18
langchain[anthropic]==0.3.18
langchain[google-vertexai]==0.3.18
langchain[aws]==0.3.18
langchain[cohere]==0.3.18
langchain[fireworks]==0.3.18
langchain[mistralai]==0.3.18
langchain[together]==0.3.18
langchain-google-genai==2.0.9
langchain-nvidia-ai-endpoints==0.3.9
langchain-ibm==0.3.6
databricks-langchain==0.3.0
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-ai21==1.0.1
# langchain-upstage==0.6.0  # langchain-upstage==0.7.3  # FIXME: temporarily disabled since causes errors
langchain-xai==0.2.0

text-generation==0.7.0
google-search-results==2.4.2
numexpr==2.10.2
langchainhub==0.1.21
sentencepiece==0.2.0
jinja2==3.1.5
accelerate==1.3.0
ollama==0.4.7
python-box==7.3.2

https://codeberg.org/poezio/slixmpp/archive/slix-1.9.0.tar.gz

# MongoDB driver for Python
motor==3.7.1
pymongo==4.13.0

papermill[all]==2.6.0
scrapbook[all]==0.5.0
nbconvert==7.16.6
nbformat==5.10.4
jupyter-client==8.6.3

python-pptx==1.0.2

pdf2image==1.17.0
pytesseract==0.3.13

# torch==2.6.0
# torchvision==0.21.0
# https://github.com/facebookresearch/detectron2/archive/refs/tags/v0.6.tar.gz

# llama-cpp-python==0.3.7
# transformers==4.48.3
# bitsandbytes==0.45.2
# langchain-openai==0.3.5
# langchain-ollama==0.2.3
# langchain-anthropic==0.3.7

qiskit==2.0.1
qiskit-ibm-runtime==0.39.0
# qiskit-aer==0.17.0
# qiskit-aer-gpu==0.15.1
# qiskit-aer-gpu-cu11==0.17.0

hvac==2.3.0

stripe==11.4.1
opensearch-py==3.0.0
prometheus-client==0.23.1
